Add collector for additional section-level metrics.
This is used to make two new graphs on the CA dashboard: exit
rates per section, and time spent per section.

In the dashboard config in Spotlight, we then specify which
section paths we want to show on the graph.

Details of the collector:

* Adding 'pagePathLevel1' to the dimensions lets us roll up the
  metrics by top-level sections of the site: so for example, we
  get metrics for /allowance/about-you and /allowance/your-details
  back as one data point, /allowance.
* We use the 'avgTimeOnPage' metric to calculate the average
  time spent within each section, and the 'exits' and 'users'
  metrics to calculate the overall exit rate within each section.

In theory we should be able to use the 'exitRate' metric to measure
exit rates, but I'm not convinced this is working at the section
level.