Add journey queries for civil claims
This service is the first to use goals in GA to count starts and
completions. This is represented differently than the usual eventLabel
and uniqueEvent count as they are metrics.

We need to use two queries because we expect a different row for the
count of starts and completes.

This uses the stage in the _id field so that the start and complete
stages do not overwrite eachother. This doesn't happen usually because
we have dimensions included in the query and these are used in the _id
generation.