Add entry for piwik.core in cronjobs.py
`cronjobs.py` expects there to be an entry in this hash in order to work
out how often it should run the collector. If there is one missing it
blows up. There wasn't one added for piwik so it blew up, adding this
should make it all run buttery smooth.

As it is similar to GA I have configured it to run at the same frequency
as it. Daily for the win!