Fix schema failings
We have now defined schemata for the collector types and these were in
breach. id is expected rather than ids that ends up getting submitted to
the API, because gapy. idMapping as always been idMapping so the id
structure of that dataset is going to change as this will finally work.