Allow credentials to be picked using regexp
We want to add some credentials that match part of the data_group, so
being able to define a pattern rather than the whole string would be far
more concise. Being able to default a set is also a requirement of the
shift to breaking out credentials to per department credentials.