Use correct token files for collectors
During the collector migration, we wrongly assumed that the token used
for the collector always matched the slug of the collector type. This
caused all govuk content, piwik and webtrends collectors to break as
they do not follow this pattern. So we have to add some special cases
for these.

This code is a bit ugly, but once the collectors are running fully
through stagecraft they will be able to derive the token used to write
to the dataset from the dataset config within stagecraft, so this is
only temporary.

This was tested by generating the crontab before and after the changes
were made, and checking the only collectors which changed token were the
broken ones.